app-ecommerce/src/main/java/br/com/itau/ecommerce/service/impl/CartServiceImpl.java
package br.com.itau.ecommerce.service.impl;


import br.com.itau.ecommerce.model.Cart;
import br.com.itau.ecommerce.model.OrderMain;
import br.com.itau.ecommerce.model.ProductInOrder;
import br.com.itau.ecommerce.model.User;
import br.com.itau.ecommerce.repository.CartRepository;
import br.com.itau.ecommerce.repository.OrderRepository;
import br.com.itau.ecommerce.repository.ProductInOrderRepository;
import br.com.itau.ecommerce.repository.UserRepository;
import br.com.itau.ecommerce.service.CartService;
import br.com.itau.ecommerce.service.ProductService;
import br.com.itau.ecommerce.service.UserService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.Collection;
import java.util.Optional;
import java.util.Set;

/**
 * Autor Lucas Zampoli 20/05/2020.
 */
@Service
public class CartServiceImpl implements CartService {
    @Autowired
    ProductService productService;
    @Autowired
    OrderRepository orderRepository;
    @Autowired
    UserRepository userRepository;

    @Autowired
    ProductInOrderRepository productInOrderRepository;
    @Autowired
    CartRepository cartRepository;
    @Autowired
    UserService userService;

    @Override
    public Cart getCart(User user) {
        return user.getCart();
    }

    @Override
    @Transactional
    public void mergeLocalCart(Collection<ProductInOrder> productInOrders, User user) {
        Cart finalCart = user.getCart();
        productInOrders.forEach(productInOrder -> {
            Set<ProductInOrder> set = finalCart.getProducts();
            Optional<ProductInOrder> old = set.stream().filter(e -> e.getProductId().equals(productInOrder.getProductId())).findFirst();
            ProductInOrder prod;
            if (old.isPresent()) {
                prod = old.get();
                prod.setCount(productInOrder.getCount() + prod.getCount());
            } else {
                prod = productInOrder;
                prod.setCart(finalCart);
                finalCart.getProducts().add(prod);
            }
            productInOrderRepository.save(prod);
        });
        cartRepository.save(finalCart);

    }

    @Override
    @Transactional
    public void delete(String itemId, User user) {
        var op = user.getCart().getProducts().stream().filter(e -> itemId.equals(e.getProductId())).findFirst();
        op.ifPresent(productInOrder -> {
            productInOrder.setCart(null);
            productInOrderRepository.deleteById(productInOrder.getId());
        });
    }



    @Override
    @Transactional
    public void checkout(User user) {
        // cria o pedido
        OrderMain order = new OrderMain(user);
        orderRepository.save(order);

        //
        //limpar pedido e diminuir o estoque
        user.getCart().getProducts().forEach(productInOrder -> {
            productInOrder.setCart(null);
            productInOrder.setOrderMain(order);
            productService.decreaseStock(productInOrder.getProductId(), productInOrder.getCount());
            productInOrderRepository.save(productInOrder);
        });

    }
}
